LOG4PHP-81 Boolean values in .properties files and incorrectly parsed by LoggerConfiguratorIni.
Currently boolean values in log4php.properties have to be written as "true" and
"false", i.e. as strings because they do not get parsed correctly else.

This is quite irritating as the ini file format as well as the name of the
configurator suggests that this file does behave like a PHP .ini file where
true/false/on/off etc. can be used.  Moreover it is no problem with log4j and
the properties files should be interchangeable if possible. 

The patch basically it does three things:

1. LoggerOptionConvert::findAndSubst() no longer tries to be clever and skips
   settings where empty($value)==true because PHPs parse_ini_file() already
   converts a false without quotes to an empty string.

2. As the generic options parser cannot distinguish if the user has written the
   number 0 or the boolean false (both is converted to string(1)"0" by
   parse_ini_file()) etc. the conversion has to be done in the setter methods
   because only they know if they want numbers or booleans for their attributes.

3. Those setter methods can use the already existing toBoolean() method.
   Although until now it did not handle all cases that parse_ini_file() mangles
   correctly.

A phpunit test case is also attached.



git-svn-id: https://svn.apache.org/repos/asf/incubator/log4php/trunk@822461 13f79535-47bb-0310-9956-ffa450edef68