LOG4PHP-144: LoggerAppenderMail should set Content-type header
Modified appender to set the Content-type header which it retrieves
from the layout.

Also improved tests, and removed the dry mode testing. It's better to get
the appender to actually send an email. For this I used the
log4php@gmail.com address which is accessible by the Logging Services PMC.